+++ saritasa_drf_tools/filters/ordering.py
@@ -1,12 +1,86 @@
 import collections.abc
 import typing
+import warnings
 
+from django.conf import settings
 from django.core import exceptions
-from rest_framework import filters, views
+from django.db import models
+from rest_framework import filters, request, views
 
 
 class OrderingFilterBackend(filters.OrderingFilter):
-    """Custom OrderingFilter for better support of openapi."""
+    """Custom OrderingFilter with additional features.
+
+    Provides:
+    - Better support of openapi via drf-spectacular.
+    - Extra kwargs for ordering fields.
+
+    """
+
+    def get_ordering(
+        self,
+        request: request.Request,
+        queryset: models.QuerySet,
+        view: views.APIView,
+    ) -> collections.abc.Sequence[models.OrderBy]:
+        """Adjust ordering params.
+
+        Wrap ordering fields in `models.OrderBy` and pass extra ordering kwargs
+
+        """
+        ordering = set(
+            super().get_ordering(
+                request,
+                queryset,
+                view,
+            )
+            or (),
+        )
+
+        is_null_first = getattr(
+            settings,
+            "SARITASA_DRF_ORDERING_IS_NULL_FIRST",
+            None,
+        )
+        is_null_last = getattr(
+            settings,
+            "SARITASA_DRF_ORDERING_IS_NULL_LAST",
+            None,
+        )
+
+        ordering_fields_extra_kwargs = getattr(
+            view,
+            "ordering_fields_extra_kwargs",
+            {},
+        )
+        fields_from_extra_kwargs = set(ordering_fields_extra_kwargs.keys())
+        if unknown_fields := fields_from_extra_kwargs - ordering:
+            view_action = getattr(view, "action", "list")
+            warnings.warn(
+                f"Unknown ordering fields: {','.join(unknown_fields)}"
+                f" defined(ordering_fields_extra_kwargs) in {self.__class__}"
+                f"({view_action}).",
+                stacklevel=2,
+            )
+        adjusted_ordering: list[models.OrderBy] = []
+        for order_field in ordering:
+            cleared_order_field = order_field.removeprefix("-")
+            order_by_kwargs = {
+                "descending": order_field.startswith("-"),
+                "nulls_first": is_null_first,
+                "nulls_last": is_null_last,
+            }
+            order_by_kwargs.update(
+                ordering_fields_extra_kwargs.get(cleared_order_field, {}),
+            )
+            adjusted_ordering.append(
+                models.OrderBy(
+                    expression=models.F(cleared_order_field),
+                    **order_by_kwargs,
+                ),
+            )
+
+        return adjusted_ordering
 
     def get_schema_operation_parameters(
         self,