+++ CHANGELOG.md
@@ -6,6 +6,8 @@ We follow [Semantic Versions](https://semver.org/).
 
 - Make `OrderingFilterBackend`/`SearchFilterBackend` show `action` with openapi warnings
 - Add `NestedFieldsMixin` for serializers
+- Add ability to add extra kwargs for ordering fields in `OrderingFilterBackend`.
+Also add new settings `SARITASA_DRF_ORDERING_IS_NULL_FIRST` and `SARITASA_DRF_ORDERING_IS_NULL_LAST`.
 
 ## 0.1.0